Update frontend to dynamically display work experience
Played about with Angular routing and resolving but couldn't get it to work, this is due to resolves not working properly for the root component. Rather than mess about with this even more, I've decided to wrap the main container in an ngIf so that it will not render/display until data is loaded.

I've stashed my previous attempt in the folder 'old' as it may come in useful in the future.