Update sidebar to be dynamically populated (except for PGP field)
We should at a future point (sooner rather than later), re-evaluate how we are handling entities so as to simplify this. Seems like the site and profile entities could potentially be merged and we may not need to create 'basic' versions of occupation and profile, as seen in profile-basic.ts.